Clean the project by defining modules.
For the hole system_monitoring thing, this commit is defining a new
module and a module for values as well.
I have decided that the ValueType Enum should be private to make it
easier to change the inner behaviour. There I have to define new
functions, these functios call simple and percent and these functions
are creating the actual value.